classes form 'game.rb' were relocated into separate files
