+++ core/bj_exception.rb
@@ -1,0 +1,1 @@
+class BJException < RuntimeError; end
