+++ core/game.rb
@@ -1,45 +1,7 @@
 require_relative 'player'
 require_relative 'deck'
-
-class GameState
-  attr_reader :dealer_cards, :player_cards, :player_score, :dealer_score
-
-  def initialize(dealer_cards, player_cards)
-    @dealer_cards = dealer_cards
-    @player_cards = player_cards
-    @player_score = assess(@player_cards)
-    @dealer_score = assess(@dealer_cards)
-  end
-
-  DEFAULT_TJQK_VALUE = 10
-  WIN_POINTS = 21
-  ACE_MAX = 11
-  ACE_MIN = 1
-  ACE_DIF = ACE_MAX - ACE_MIN
-
-  private
-
-  def assess(cards)
-    aces = 0
-    total = cards.inject(0) do |res, card|
-      case card.name
-      when '*' then '???'
-      when /[K,T,Q,J]/ then res + DEFAULT_TJQK_VALUE
-      when /A/
-        aces += 1
-        res + ACE_MAX
-      when /[2-9]/ then res + card.name.to_i
-      else raise BJException, 'Incorrect Card name'
-      end
-    end
-
-    total.is_a?(Integer) && total > WIN_POINTS ? true_aces_value(total, aces) : total
-  end
-
-  def true_aces_value(total, aces)
-    [((total - WIN_POINTS + ACE_DIF - 1) / ACE_DIF), aces].min * ACE_DIF
-  end
-end
+require_relative 'bj_exception'
+require_relative 'game_state'
 
 class Game
   def initialize(player_name = 'New Player')
@@ -92,9 +54,7 @@ class Game
   end
 
   def dealer_turn
-    if internal_state.dealer_score < 17 && @dealer.hand.size < 3
-      @dealer.add_card(@deck.take_card!)
-    end
+    @dealer.add_card(@deck.take_card!) if internal_state.dealer_score < 17 && @dealer.hand.size < 3
 
     check_final!
   end
@@ -104,5 +64,3 @@ class Game
       @player.hand.size == 3 && @dealer.hand.size == 3
   end
 end
-
-class BJException < RuntimeError; end
